Removed containerNameOverride
Removed recently added containerNameOverride from container mapping. Since container name is not used during service deployment in EC-Docker, just the service name, the override will only need to be supported at the service level.
The serviceNameOverride is expected to be wired in soon, it is just a place-holder for now.