Removed certificates copied on the agent by the procedure step
o Set the files and directory where the certificates are copied to delete on exit
o Cleaned up the logic in the procedures to use DockerClient for managing the certificates.
o Refactored check cluster connection logic for both 'Check Cluster' procedure and 'test connection' in 'CreateConfiguration'
o Bumped up the grapes artifact version. The dependent jars were updated in the last commit.