Updates made while trying out the service catalog item for Docker Compose
o Made the EFClient consistent with the version in EC-Kubernetes for the most part.
o Refactored the port handling for docker service and container.
o Renamed the parameters to remove the 'ec_' which is currently supressed by the UI so the proc cannot work with these names in the service catalog UI.